Waitlist retries
