+++ kifi-backend/modules/shoebox/app/com/keepit/controllers/core/AuthController.scala
@@ -606,102 +606,6 @@ class AuthController @Inject() (
       DiscardingCookie(Authenticator.cookieName, Authenticator.cookiePath, Authenticator.cookieDomain, Authenticator.cookieSecure))
   }
 
-  def requestToTwitterWaitlist = MaybeUserAction { implicit request =>
-    val session = request.session
-    request match {
-      case requestNonUser: NonUserRequest[_] =>
-        Redirect("/signup/twitter?intent=waitlist").withSession(session + (SecureSocial.OriginalUrlKey -> "/twitter/thanks"))
-      case ur: UserRequest[_] =>
-        val twitterSocialForUser = db.readOnlyMaster { implicit s =>
-          socialRepo.getByUser(ur.user.id.get).find(_.networkType == SocialNetworks.TWITTER)
-        }
-        if (twitterSocialForUser.isDefined) {
-          // if user already has twitter account, go to thanks
-          Redirect("/twitter/thanks")
-        } else {
-          // if user does not have twitter account, link it!
-          Redirect("/link/twitter").withSession(session + (SecureSocial.OriginalUrlKey -> "/twitter/thanks"))
-        }
-    }
-  }
-
-  private def pollDbForTwitterHandle(userId: Id[User], iterations: Int): Future[Option[String]] = {
-    /*
-     * This is not good code, but appears to be necessary because initially, when a social
-     * account is brought in, the SocialUserInfo record is not complete. Namely, it's missing
-     * the profileUrl field. Only after our system syncs with the social network do we get this.
-     * Unfortunately, in this code path, it's nearly always too fast, so it's empty. However,
-     * we need the handle here. So we poll the db every 300 ms (up to `iterations` times) looking for it.
-     * We aggressively log if we can't get it fast enough.
-     *
-     * What we should do:
-     *   1) Handle creation of SocialUserInfo records ourselves, filling them to our heart's desire.
-     *   2) Minimize dependency on SecureSocial's static approach. It won't work much longer.
-     */
-    def checkStatusOfTwitterUser() = {
-      db.readOnlyMaster { implicit session =>
-        socialRepo.getByUser(userId).find(_.networkType == SocialNetworks.TWITTER).flatMap { tsui =>
-          if (tsui.state == SocialUserInfoStates.CREATED) {
-            log.info(s"[checkStatusOfTwitterUser] Still waiting on ${tsui.networkType}/${tsui.socialId}")
-            None // pending sync, keep polling
-          } else if (tsui.state == SocialUserInfoStates.FETCHED_USING_SELF && tsui.getProfileUrl.isDefined) { // done
-            log.info(s"[checkStatusOfTwitterUser] Got it for ${tsui.networkType}/${tsui.socialId}")
-            Some(Right(tsui.getProfileUrl.map(url => url.substring(url.lastIndexOf('/') + 1)).get))
-          } else { // other
-            log.info(s"[checkStatusOfTwitterUser] Couldn't get handle of ${tsui.networkType}/${tsui.socialId}")
-            Some(Left(()))
-          }
-        }
-      }
-    }
-
-    var times = 0
-    def timeoutF = play.api.libs.concurrent.Promise.timeout(None, 600)
-    def pollCheck(): Future[Option[String]] = {
-      timeoutF.flatMap { _ =>
-        checkStatusOfTwitterUser() match {
-          case None if times < iterations =>
-            times += 1
-            pollCheck()
-          case None | Some(Left(_)) => // Timed out or weird error
-            Future.successful(None)
-          case Some(Right(res)) => // Got handle!
-            Future.successful(Some(res))
-        }
-      }
-    }
-    pollCheck()
-  }
-
-  def thanksForTwitterWaitlist = MaybeUserAction { implicit request =>
-    val session = request.session
-    request match {
-      case requestNonUser: NonUserRequest[_] =>
-        Redirect("/twitter/request")
-      case ur: UserRequest[_] =>
-        val noTwitter = db.readOnlyMaster { implicit session =>
-          socialRepo.getByUser(ur.userId).find { s =>
-            s.networkType == SocialNetworks.TWITTER &&
-              (s.state == SocialUserInfoStates.FETCHED_USING_SELF ||
-                s.state == SocialUserInfoStates.CREATED)
-          }.isEmpty
-        }
-        if (noTwitter) {
-          Redirect("/link/twitter?intent=waitlist").withSession(session + (SecureSocial.OriginalUrlKey -> "/twitter/thanks"))
-        } else {
-          pollDbForTwitterHandle(ur.userId, iterations = 50).map { twRes =>
-            twRes match {
-              case Some(handle) =>
-                twitterWaitlistCommander.addEntry(ur.userId, handle)
-              case None => // we failed :(
-                airbrakeNotifier.notify(s"Couldn't get twitter handle in time. userId: ${ur.userId.id}. They want to be waitlisted.")
-            }
-          }
-          MarketingSiteRouter.marketingSite("twitter-confirmation")
-        }
-    }
-  }
-
   // New signup pages
 
   // todo, this is signup
