+++ kifi-backend/modules/shoebox/conf/shoeboxService.routes
@@ -91,8 +91,8 @@ GET     /get                        @com.keepit.controllers.website.HomeControll
 
 GET     /twitter                    @com.keepit.controllers.website.TwitterWaitlistController.twitterWaitlistLanding()
 GET     /site/twitterQueuePosition  @com.keepit.controllers.website.TwitterWaitlistController.getFakeWaitlistPosition()
-GET     /twitter/request            @com.keepit.controllers.core.AuthController.requestToTwitterWaitlist()
-GET     /twitter/thanks             @com.keepit.controllers.core.AuthController.thanksForTwitterWaitlist()
+GET     /twitter/request            @com.keepit.controllers.website.TwitterWaitlistController.requestToTwitterWaitlist()
+GET     /twitter/thanks             @com.keepit.controllers.website.TwitterWaitlistController.thanksForTwitterWaitlist()
 
 
 POST    /sms                        @com.keepit.controllers.website.HomeController.sendSmsToGetKifi()
